cleanly separate the UTM coordinates from the Rock convention's NWU coordinates
This changes method names, and also changes the behaviour of the convert
methods in case the solution is a NO_SOLUTION. RBS can cleanly represent
NO_SOLUTION, so simply return a RBS in this case. The caller can still
manually check for NO_SOLUTION if he wants a different behaviour.